refactor(pdf): isolate source crop asset validation
